[HSG] Allow class directive without policy
In some cases, we want to add a policy for a member (field/method) in
the policy file for a class that already has an annotation.

Previously, we had to always have some policy in the class line, which
could cause confusion because this will essentially override the
annotation, so even when you change the class annotation, that wouldn't
be picked up.

Now update the parser so that a class directive is allowed to omit
a policy.

Also fix b/388562869. But I still need to ignore certain lines in diff
because the merged golden files are already missing the expected lines.

Flag: EXEMPT host test change only
Bug: 292141694
Bug: 388562869
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh -s
Change-Id: Ifa37cd0ff74026bc1ba2c7e8599b159fa0a07412

Former-commit-id: aeac86ed5d218685e163d285d0f82f5b5f224f85