[SB][Notif] Ensure promoted notifications are never sticky.
This is done by making sure `isSticky()` returns false if the
notification is promoted. This means promoted notifications won't be
sticky if you tap the status bar chip to show the HUN, or if an app
posting a promoted notification makes it alert.

Note: In practice, I think this case isn't possible because `mExpanded`
should only be set to true when the expansion caret is tapped and RONs
shouldn't show the expansion caret. But this adds protection just in
case.

Fixes: 385727186
Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips

Test: Post two RONs, tap on one status bar chip to show HUN, then tap
expansion caret on the HUN. Then, tap the other status bar chip ->
verify HUN for the other chip shows
Test: atest HeadsUpManagerImplTest

Change-Id: Ibce9cb64125deeb00fff532033ea73c31c5a8654

Former-commit-id: 77fb9ee03d9374e7a7f0ca5b701520d8a2278319