Use correct root leash if PiP is the only target
If PiP is the only target in a transition, root leash's z-index
might end up behind that of PiP's ancestor's top child.
For example, if PiP is moved to back in a transition its last layer
if behind that of the top task, and that is the layer used when setting
the z-index of the new transition root leash.

Instead, since PiP is an always-on-top task, we wanna make sure the root
leash is on top throughout the transition when an initially pinned task
is a target.

Bug: 389767099
Flag: com.android.wm.shell.enable_pip2
Test: manually repro the steps in the bug
Change-Id: I88c2de682b8e606f1aa815dc7bc3d8ad513601df

Former-commit-id: 4d5fa9fd2aa61d8e8f1b444e77c1fb3edcbca84f