Set IME Switcher nav hint based on IME visibility
The navigationIconHints are computed in
Utilities#calculateNavigationIconHints, using the given value of
showImeSwitcher. This is modified before the call to be true only
if showIme is also true.

This change moves the modification to inside the method, to ensure
consistency. Note, it wouldn't make sense to request showing the IME
Switcher button when the IME itself is not visible.

Flag: EXEMPT refactor
Bug: 366129400
Test: atest NavigationBarTest#testSetImeWindowStatusSysuiState_ImeNotVisibleImeSwitcherButtonVisible
Change-Id: Ia6a04ecd3700acf2078ae2cb4cd2412bcd7caa51

Former-commit-id: 43ce1d62e18b1a67227dc70880ce06eb70069e0f