Use blurred wallpaper for Glanceable Hub background.
This changelist adds support for glanceable hub transitions to influence
the blurred properties of the wallpaper. A new interface,
GlanceableHubTransition, defines a flow that exposes the amount of blur
to be applied. Implementations of these interfaces, found in the
existing GlanceableHub transition view models, are observed by the
WindowRootView logic, combining it with other sources of blur, such as
bouncer and shade.

This changelist also introduces GlanceableHubBlurProvider, which helps
maintain a consistent blur experience across the transitions.

Test: updated TransitionViewModelTests
Bug: 389788272
Flag: com.android.systemui.glanceable_hub_blurred_background
Change-Id: Ie04f04397f1d40c96377a1a91a1b5b816e7f42b1

Former-commit-id: 762b011d86bb6855d3329d9e22ccbce66be182e3