Hide NSSL while notification rebinding is in progress after display change
After the shade window changes display, notifications are re-inflated.
While most of the times this is fast enough and non-noticable, in case of many notifications of throttled CPUs this takes time (cases of >400ms have been recorded).

Inflation is done in the background, and views with the old dimensions (correct for the source display, but not the destination one), might be visible for some frames until the new inflations succeed.

With this change, the notification stack is hidden immediately after the resources are updated to the new configuration (associated with the new display). This means that while inflations are going on in the background and views change configuration, the flicker is not user-visible, and the UI thread is not blocked either (so, the shade can still expand jank-free).

Bug: 362719719
Bug: 390371121
Test: ExpandableNotificationRowControllerTest, ShadeDisplaysInteractorTest, ExpandableNotificationRowControllerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I4b3d32cb108e506b917696a9091a57d774f8aaab

Former-commit-id: 988f68315318fce316604cdd74c54932d5d793b6