Don't keep dream alive under hub with hub v2
With hub v2, there is no longer a swipe to/from the hub to the dream so
there's no need to keep the dream alive. While this is not directly
needed for hub v2, this improves the hub's performance and reduces the
complexity of the hub interaction model and shade/bouncer/scrim
handling logic.

Bug: 389108560
Test: atest CommunalDreamStartableTest DreamOverlayServiceTest
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I59ce39214a8471526abb929e10767e16eb1da75c

Former-commit-id: 471e5a310ab9f3f688d6e7181efa0a6d096ba42a