Handle launch adjacent resize
* Re-use path for enter transition that dragging an app to replace
an app already in split uses for resizing.
* Ideally we can resize after start animation or add the changes we
want in the WCT we return in handleRequest(). Currently that work
is done in SplitLayout#flingDividerToCenter() which starts its own
transition for resizing after animating to center. In our case I think
we want to set the final state and then animate so it plays well with
handleRequest() and then startAnimation()?
* Technically just calling flingDividerToCenter() here works but
there is too much flickering in the UI during resizing, not sure if the
resize surface updates are interfereing w/ the app that is starting?

Test: Doesn't resize when the task being replaced is pippable since
we need to let MixedHandler handle that. But resizes for other cases.
Bug: 383624769
Flag: EXEMPT bug fix

Change-Id: Iccd3398900089177928cd51fab999a421610ab45

Former-commit-id: 5ccfe4e30c7e294209d21e6d447abcc2d42f1ac2