Allow starting activity from visible home if app switch stopped
After returning to home, stopAppSwitches will be called to set
APP_SWITCH_DISALLOW for 500ms. Then it falls to the condition
mCallingUidHasNonAppVisibleWindow because home has a visible
Taskbar window. But it will be logged as wtf event for
BAL_ALLOW_NON_APP_VISIBLE_WINDOW, which is noisy.

This aligns the same condition as
checkBackgroundActivityStartAllowedByRealCallerInForeground.

Bug: 291794451
Flag: EXEMPT bugfix
Test: Swipe to home and launch an app within in a short time.
      There should not have a wtf log for home.
Change-Id: I6ba31b324dc9a3dfe7f205946b9bdb8d7bd5d41b



Former-commit-id: 3e45f122590db702b53d645502e6c2d1a852855e