Always call getOrCreatePackagePreferences in getNotificationChannels
Once flag is on, we always call getOrCreatePackagePreferences when an app calls getNotificationChannels. This affects the overall behavior for a very small slice of apps (ones that qualify for shouldHaveDefaultChannel) in the specific time before the app has interacted with notification settings at all.

Bug: 381131846
Test: NotificationManagerTest, PreferencesHelperTest, presubmit
Flag: android.app.nm_binder_perf_cache_channels
Change-Id: Ie6e13edcf0c164feed8f5cf4dcd656ede82171e7

Former-commit-id: c41dcd7b8ee121b2263a6b3aedd299bcd5c0f8b7