Simplify behavior of TestLooperManager#isBlockedOnSyncBarrier.
Previously isBlockedOnSyncBarrier would only return true if
the corresponding queue had a sync barrier as the next message,
AND a another message was waiting to be executed.

This is problematic for espresso (and I'd argue non-intuitive), which uses
isBlockedOnSyncBarrier to determine if queue is idle or not. The presence
of a sync barrier should make the queue idle.

This commit simplifies the definition and implementation of
isBlockedOnSyncBarrier to strictly mean 'the next message in the
queue is a sync barrier'.

Flag: android.os.message_queue_testability
Test: atest TestLooperManagerTest
Bug: 112000181
Change-Id: I443d3c8ed8d439f058136d55f70c9b8c2c444149

Former-commit-id: d96404938b08d25c553681dd8b5c30aa872d259b