Update smartspace visibility to correct target during clock size transition
Sometimes ClockSizeTransition is given the wrong target visibility for
smartspace date/weather view. This causes it override the correct value
when it comes in later. To correct this, we ensure we're animating to
the correct target value every frame.

This may cause animation hiccups in some edge-cases, including the case
that's being fixed, but it corrected fast enough that it wasn't visible
in my testing.

Bug: 376075102
Flag: NONE Bugfix
Test: Manually checked reproduction
Change-Id: I5bfd754c82c15759348a2ecf8ba94a133f2d547b

Former-commit-id: 7fe71e59ed00a9da86c5b59dc9c2428b69199529