Separate the notion of "note bubble" and "app bubble"
There is some very specific behavior associated with notes bubbles so
we need a separate method to check for this. This CL does this by
specifying a unique key for notes bubbles & setting a flag on the bubble.

Updates uses of getting the key / note bubble creation to the new methods.

Notes bubbles & apps bubbles have some shared behavior so a notes bubble
is still considered an app bubble.

Flag: EXEMPT trivial refactor
Test: atest SystemUITests
Test: atest BubbleControllerTest (in other CL)
Bug: 390496034
Change-Id: I4d921f5a048eec520fe2eb5d39928a8934c9e024

Former-commit-id: 0c3c3d9d16e87c4e2ca3f43f7143064eb964d619