Allow showing hub immediately on view load
This change makes it possible to show the hub immediately upon load
from a state where keyguard is not showing and the hub container view
is not loaded. This is needed to support pressing power button to show
the hub from launcher (while postured & charging).

Bug: 378170118
Test: atest CommunalSceneRepositoryImplTest
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I5a85e15270af3efebd1b4157707f03700d4c4f9d

Former-commit-id: b3d5a84ac5bf4c3a346a1d266c425053b33346cc