Fix the visibility threshold of BaseContentOverscrollEffect
This CL fixes the visibility threshold used when animating the offset of
any BaseContentOverscrollEffect. For some reason, the previous threshold
that was passed in the Animatable constructor was ignored and we have to
provide it in the spec instead.

I also decided to increase the threshold to 1f, as there was no
noticeable visual difference between 1f and 0.5f but there was a
noticeable latency difference.

Bug: 378470603
Test: Manual, tried with a very high threshold and noticed that it
 wasn't picked up but now it is.
Flag: com.android.systemui.scene_container
Change-Id: If913aeb52f130cbbc171b2b6ea531559f7b4e006

Former-commit-id: 1331d7abcd80ed5d5b1994e73ee2f1d84cb90486