Don't listen to UDFPS if communal hub is showing
When the communal hub is showing, we should not listen to UDFPS as there
are visual widget elements above the sensor. When communal exits, we
should re-enable the sensor. This also ensures that if the alternate
bouncer is shown above the communal hub, we also re-enable the sensor.

Fixes: 385762852
Test: atest KeyguardUpdateMonitorTest
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I1bdb81672857c61245a6df430127af363b961bc0

Former-commit-id: 244bcaed5cd2352b9d7947ae5c04d75aa8a2b592