Distinguish the reason why decoding failed.
We were throwing a RuntimeException that the message hash wasn't in the
viewer config file, but this can also happen when the viewer config file
wasn't loaded into memory from disk, which is more likely a bug.

Update the exception to clarify what's happening.

Test: atest com.android.internal.protolog.ProtoLogViewerConfigReaderTest
Bug: 388235335
Flag: EXEMPT updating error message clarity
Change-Id: I484c918ec196d57bb910900e3a7a7870923d64a5

Former-commit-id: 8623733336a2a73d98f5107d0942d3db40c612dc