Introduce bubble drag indicators
Add bubble left and right drag indicators.
Define the drag areas that are used for bubbles.
The visual indicator is a placeholder. We need to update the size based
on the spec.
Moves the app to a bubble using the existing animation. Which needs to
be improved to handle resize from the drag state.

Bug: 388851898
Test: atest WMShellUnitTests:DesktopModeVisualIndicatorTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: Idf8dd7d16a9c0d8b42ec4eba94059384d07c3234

Former-commit-id: 884afc81c97205cfbbb00dda8b860ee497dfb5df