Acquire wakelock before alarm dispatch
This is required to avoid stalls in the app if a suspend operation is
already under-way for some reason. The expectation is that the userspace
wakelock acquisition will either block or succeed, in both cases
guaranteeing that the app executes continously while processing the
alarm.

Test: atest FrameworksMockingServicesTests:com.android.server.alarm

Flag: com.android.server.alarm.acquire_wakelock_before_send

Bug: 391413964
Change-Id: I56a0739a60aa58139d40e8d64523328d5430b6ec

Former-commit-id: 6714f64fdafe6bd29b8714ad23509080b1ac1355