[PiP2 on Desktop] Drag-corner-to-resize PiP.
Recall: http://recall/clips/4bb9825b-e1a9-4971-8f30-1f78229e8282

Implement drag-corner-to-resize PiP for PiP2 on Desktop Mode. For now,
this is only enabled if the PiP was entered while in Desktop Mode
(i.e. minimized into PiP).

The implementation for drag-corner-to-resize is largely taken from the
PiP1 implementation in If639834e377a705dfbe8d6eee7bda988b61edb32.

Test: Manually check that PiP can be resized via dragging the corners
when it is entered while in Desktop Mode, and not if it is entered in
fullscreen.
Bug: 378744674
Flag: com.android.window.flags.enable_desktop_windowing_pip

Change-Id: Ief6b24144578d54a910b74660d9d5a30756b0c57

Former-commit-id: 96b677267a871ca7999d0a82e042918909e0512b