Ensure CallerInfoAsyncQuery uses HSUM aware method to check user id of current process.
CallerInfoAsyncQuery#getCurrentProfileContentResolver was using a deprecated non HSUM aware method in order to check the user id of the current process. This was causing some contacts provider UT to fail when the device is HSUM mode. This CL updates CallerInfoAsyncQuery to use the HSUM aware reccomended method UserHandle#myUserId instead.

Test: atest CallerInfoIntegrationTest#testCallerInfo
Bug: 383209047
Flag: EXEMPT bugfix
Change-Id: I9032c2781c2502848d246be3cf83492a5a0d7228

Former-commit-id: 776b1cd5cf93aa0ec4bfad137e846ec0c145c056