Increase the number of system PIC nonces
Increase the maximum number of system PIC nonces from 64 to 128. A
recent image used 45 nonces, which is uncomfortably close to 64.

The number of bytes is left at 8192 since the same recent image only
used 976 bytes.

A test was added to verify that the running configuration values and
to verify that the maximum number of nonces that can be created
matches the configuration.

An unused attribute was deleted fron NonceStore.

Flag: EXEMPT bug-fix
Bug: 392920391
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
Change-Id: I48fdd50dc430c58a798acec43ce7651392a54868

Former-commit-id: c409163fbe1d96c32ba7fc735508016e444212ad