[Notif redesign] Align messaging content under app icon
For most templates, the content needs to come directly under the top
line text in the expanded view, so we calculate a special margin to
align things correctly. For messaging notifications however, since they
display the avatar on the left side below the app icon, we need to leave
enough space between the header with the icon and the actual content to
avoid the icons overlapping.

Bug: 378660052
Test: visual test, screenshot tests to come later
Flag: android.app.notifications_redesign_templates
Change-Id: Iae9bfcd4de08014660bbb91c6ca7e28866c842e5

Former-commit-id: fbcb8acb3af4277f16afae63d2da1c07e9e61b43