Translate digits on AOD to not be quite so wide
This uses two views stacked vertically instead of 4 individual views.
Doing this improves behavior for moving from very wide axes to relative
thin ones on AOD, and provides much better spacing.

I'm hopeful this will also improve our memory usage as it effectively
halves the number of text animators and views we're allocating in
using. We'll see about this when the metrics come in. If it is
significant, I'll probably move to one view with all four digits, but
that layout proved tricker to implement than this for the time being.

Finally, this also breaks the stepping animation. That'll need to be
addressed, but the implementation of that will depend on where we
land for number of child views we're using here. b/393577936
tracks that follow up work.

Bug: 391858439
Bug: 385338722
Test: Manually checked spacing
Flag: com.android.systemui.shared.clock_reactive_variants
Change-Id: Ib6f67763e77a3ceeeea4a5cbf97dcad6307867fe

Former-commit-id: 552c1df23e8a38d4226695d6cfaa659b783008e7