Improve some common/ui/ utilities
This CL adds the @ReadyOnlyComposable to ContentDescription.load() and
Text.load(). This should slightly decrease the number of groups
generated by the Compose compiler, which is usually better for
performance.

This CL also makes use of rememberDrawablePainter() in Icon(), when the
passed icon is an Icon.Loaded(Drawable). I haven't measured this, but
this should be much better than converting the Drawable into a Bitmap
everytime Icon is recomposed.

Bug: 372412931
Test: Manual, volume panel (only user of Icon) still worked as expected
Flag: EXEMPT performance fix
Change-Id: I56f38a562f002645f6a625cc2cdf1e96e97ae5c6

Former-commit-id: 1eb4dea0635a0e8dcbc4d2122c116b21f06d4792