[Notif redesign] Fix broken expand animation in minimized groups
With the recent change to perform a translation on the expand button for
the group header while it's being expanded via dragging, an issue was
introduced where minimized group headers would show two expanders
overlapping and moving in a weird way on expansion.

To fix this, we just need to make sure that we're not applying the
translations to the normal group header when we're actually showing the
minimized one.

Also simplified the calculateDesiredHeader method a little bit.

Bug: 378660052
Test: post low priority group and expand it via dragging
Flag: android.app.notifications_redesign_templates
Change-Id: If33c503a3ed528ea69f79c806d55e8c1a5cd89f0

Former-commit-id: 8803136f920a61b482fd22c336c96249f9b6f71a