Adding magnetic behavior for non-dismissibles and other ExpandableViews.
The behavior on these cases applies a reduction to magnetic pulling, as
well as slightly stronger haptics to indicate the increased level of
resistance.

When non-dissmisible notifications are pulled, the translation
multipliers are further reduced fot the swiped row and its neighbors.
When swiping a dismissible row, any neighbor row that is not dismissible
(including FooterView, SectionHeaderView and NotificationShelf) also
gets a reduced translation.

Test: NotificationTargetsHelperTest
Test: MagneticNotificationRowManagerImplTest
Bug: 392747453
Bug: 392738615
Flag: com.android.systemui.magnetic_notification_swipes

Change-Id: I1dac9e7c607f25fa959d5a2fff3f59aebdc81e4a

Former-commit-id: 72d225c4a977a588bf95241635387326a1e1ab97