Revert "Remove MessageQueue reflection from TestLooper"
This reverts commit ae1a3058d99cbd4636046895e848d2e937231fbf [formerly ef7309768eddd05aa6d9fcacfed9447a1a9c9584].

Reason for revert: `Build.VERSION.SDK_INT >= Build.VERSION_CODES.BAKLAVA;` is a ticking bomb for the SDK team. Can you please hold off on this assumption until we've finalized the SDK and bumped the SDK_INT for all relevant configurations (next, trunk and trunk-staging). (I bumped it to 36 for trunk-staging in ag/31522509 which triggered this code to be active and (probably) led to b/393534402)

Change-Id: I9e6279f5d4845469cd29e47501da3e8546fc71d2

Former-commit-id: c69688d86efd4eab0427be93573f0844b0966bb0