Remove drag position & size check
This CL removes a safety check that was added when starting a drag, that
was double checking that the drag position and number of pointers down
was the same that was tracked by another pointer input handler. As shown
in b/393423580 and discussed offline with the Compose team, this
assumption is not always correct depending on some very specific
timings.

Bug: 393423580
Test: None, unfortunately I was not able to reproduce given that this is
 a bug that can occur in some very specific & rare circumstances
Flag: EXEMPT crash fix
Change-Id: I3d62239ec042e5a20944f4b205383a9e44553f85

Former-commit-id: 1d38904507da9e70b33288f6b72419486f81a095