Add temporary debug logging.
This is to investigate Permission Denial: broadcast from android asks to
run as user X but is calling from uid Y. The exception is seen in many
tests using various DPM APIs, which don't result in any broadcasts
except for DEVICE_POLICY_MANAGER_STATE_CHANGED, so this seems to be the
most likely source. Normally withCleanCallingIdentity() should make it
look as if it is broadcast from uid 0.

Bug: 387259698
Test: TH
Flag: EXEMPT only logging in exceptional circumstances
Change-Id: I7242003f9c07a07e1e163fc9d32657e5781116e7

Former-commit-id: e5613ecf1d4b103af8990cd328dbe56ec5d146d5