Remove DesktopWallpaperActivity token onVanished.
Sometimes desktop wallpaper activity might be cleared up without a
transition. E.g "clear all" button on overview. When that happens the
token that we store in shell never gets removed as it was only removed
by transition observing.

This cl removes the token when onTaskVanished callback is triggerd as well.

Bug: 391501236
Test: manual
Flag: EXEMPT bugfix
Change-Id: I29be4fd67a2a6a54fc1df035ac31ca0f77483382

Former-commit-id: b2ff663cb5d3b5ca34e201ee13bfe6799a1e5ce8