Upgrade path for small/large tiles
When upgrading from a build that doesn't support dual size tiles to one
that does, we will force the current tiles for the user to be large
(once).

This is done by the following:
* If there is a list of tiles read from Settings AND
* The shared preference for large tiles is empty,

THEN we set the tiles read for that user to be large.

After that, if the user changes the tiles size or sets them to default, that
will be respected.

Test: atest com.android.systemui.qs
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Bug: 392615513
Change-Id: I3c024dc3c3f882a8a186c660d9094e6d07454c3f

Former-commit-id: b02dd916fc279fbf7a41100c87e334abb94bf30d