Make sure the current shade blur uses the old blur radius value when notificationShadeBlur flag is not enabled.
Ramping up the bouncer_ui_revamp flag led to jank in the NOTIFICATION_SHADE_EXPAND_COLLAPSE CUJ because it was using the 60dp radius by mistake.
This might help reduce the jank for now, but it might come up when we ramp up the notificationShadeBlur flag later on. Jank requires a bit more investigation, the intent of this CL to isolate the flag's impact to only the bouncer CUJs.

Bug: 394275414
Bug: 393199337
Flag: com.android.systemui.bouncer_ui_revamp
Test: NA
Test: verified manually that shade pull down still uses the old blur radius
Change-Id: I35d0f873a059f042d85903b24ca142215dd4f95e

Former-commit-id: 0a4d8d88166b77668b364efb84be6cca60804fdb