Start to bubble transition when releasing task in bubble area
When released in bubble area, signal BubbleController to start the
transition to convert the running task into a bubble.
Keep the desktop drag transition running during this.
Once the bubble transition starts, we will first attempte to merge the
animation with the running drag to desktop animation.
Detect this case in the desktop handler and cancel the ongoing drag to
desktop transition.
This will allow bubble transition handler to take over the animation and
move the task into a bubble.

Bug: 388851898
Test: atest BubbleTransitionsTestTest
Test: atest DragToDesktopTransitionHandlerTest
Test: manual, drag a fullscreen task from handle to corner, check it
  converts to bubble on release
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: Ic2b1cd096b89afdde51429f70dafea47bdd45442

Former-commit-id: 9f89e861297752d7400dc3ff21e2e771bfad2c19