Fix low light triggering on phones when entering doze
Tangor did not have doze/aod so it was not accounted for in the
conditions. While DreamManager does explicitly not override doze with
the system dream, the doze dream does not start dozing fast enough to
prevent the low light code from overriding it with the system dream.

This change adjusts the DeviceInactiveCondition to only be true if the
device is not dozing or entering dozing to prevent the possibility of
entering low light dream when unplugged and entering aod.

Bug: 394707567
Fixes: 394707567
Flag: com.android.systemui.glanceable_hub_v2
Test: atest DeviceInactiveConditionTest
      manually verified on device in dark environment
Change-Id: I83f9584940ab3e8a70125ad9dd050e77810465f4

Former-commit-id: 4f9ab76ab6c9aea09d554fc56e09e905c739114b