Fix activity's launched from notification shade's displayId
Trying to launch Settings from the settings gear on the notification
shade, or other activities like notification history, would always
launch on the default display, even if the shade is on an external
display. This is because the displayId used for launching the activity
was alwasy the default display. Now it is the display id is shade aware
since it is derived via shade aware context.

Fix: 389636334
Test: manual testing
Test: StatusBarNotificationActivityStarterTest
Flag: EXEMPT bugfix
Change-Id: I10a41cc6d6766263027766698bbc58beda2ec03e

Former-commit-id: 2adc744a3973e46bd3fa52fea9b370ce8e59bf53