Update user switching processes
Add a new callback for UserController. If the user is secure, initiate
a call to SystemUI to lockNow(), and wait for the callback. If the
callback does not arrive, crash the system after some time.

In SystemUI, add support for onBeforeUserSwitching. When this request
comes in, prepare the keyguard and show it before calling back to notify
UserController that it is ready to proceed.

Also, ensure that any switch is user that occurs after WM has been
notified that SystemUI is going away is stopped.

Bug: 322157041
Test: atest KeyguardViewMediatorTest
Flag: EXEMPT bugfix
Change-Id: I32699e32fe1dbe94af2abe7e7ade5363a2b6cb55

Former-commit-id: d8c3d450f77f77232a89ac37c9b9b266e28c0202