Animate task to bubble immediately during transition
Remove the wait for Launcher when converting a task to a bubble via
transition.
Previously we were waiting for the Launcher part to show the bubble bar
and come back to shell to show the expanded view.
This works for new bubbles that are triggered via notifications. But
adds a visible to delay when the bubble is triggered via user action
from a running app.
The delay is especially noticable when moving an app to a bubble via
drag. As the app waits on the screen before starting to move to the
bubble position.
Update the flow so that the convert to bubble animation can start
immediately when the transition is ready and the surface is created.
Add a new progress tracker class to the ConvertToBubble transition class
that can track different parts of the transition completing and then
performing the required actions.
We still have the callback from Launcher, but only use it to clear the
pending transition object from bubble. This pending transition is used
to tell other parts of the bubble code that this bubble will be expanded
via the transition animation and they do not need to animate it.

Bug: 388851898
Test: atest BubbleTransitionsTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: Ie61f67a752ed7faec728f8de6c20e779cd0e4460

Former-commit-id: d0272fd3cfe1205b021b46395a74ac1ab991167a