[flexiglass] Conditions handleShadeTouchability to run conditionally
handleShadeTouchability is responsible for the attached bug because it
sometimes emits a false right when we unlock the device, causing the
system to switch to the Lockscreen scene while needing to actually
unlock.

The power button putting the device to sleep is handled by a different
coroutine in SceneContainerStartable so that's still covered.

Fix: 395728344
Test: current tests pass
Test: manually attempted to reproduce the bug about 20 times without
success
Flag: com.android.systemui.scene_container

Change-Id: I5becb58e6b9a23a2a9f7ea6522ca12f4557c5d12

Former-commit-id: 0014623f3f9b93745b876b1392bb02dae7c5ace7