Unsuspend apps for affected user correctly
When unsuspending apps for a single user, the call was made unsuspend
apps for user 0 even when the original intent was to unsuspend apps for
user 10. Keep the target user a bit longer, and pass suspending and
target user correctly to the unsuspend call.

This caused couple of test failures on HSUM because the respective test
left all personal apps suspended, causing weird failures in subsequent
tests.

Bug: 384826303
Test: btest android.devicepolicy.cts.DeviceManagementCoexistenceTest
Test: atest SuspendPackageHelperTest
Test: atest CtsSuspendAppsTestCases
Test: atest ActivityStartInterceptorTest
Test: atest PackageManagerSettingsTests
Test: atest PackageUserStateTest
Flag: EXEMPT BUG_FIX
Change-Id: I7eb757363dee14f4cfc7c6798697cfc4ccb5a87e

Former-commit-id: efbc34726a30e9adb8ffe8bae1648031e5ce8e19