Add ACTION_DESKTOP_MODE_ENTER_APP_HANDLE_MENU latency.
This measures the delay from when the user requests to enter
desktop mode by pressing the desktop mode button in an app
handle's menu until the animation to enter desktop mode
begins.

This period is consumed by launching the home activity and
creating app headers for each desktop window, both of which
have been expensive in the past.

Flag: EXEMPT metric change
Bug: 390555563
Test: Put tangorpro in HSUM mode:
      `adb shell cmd user set-system-user-mode-emulation headless`
      Run `statsd_testdrive 306` and ensure that latency metric is
      reported. This was done with
      `enable_desktop_wallpaper_activity_for_system_user` enabled and
      disabled. The latency was close to 1 second with that flag
      disabled and ~100 milliseconds with that flag enabled. This is
      expected.
Change-Id: Ic8a1131063d16f9080bc338298cb3b497206a2f8

Former-commit-id: 28efafc2610345cbf1278b5f69c116ca2d8f1459