CastDetailsView: Introduce MediaRouteControllerContentManager.Delegate
Add a new Delegate interface to MediaRouteControllerContentManager so
that the users of this content manager can update their UI accordingly
when needed. This CL introduces 2 functions in the interface, as well as
moving the change title and button callback from the dialog to the
content manager.

This CL follows the refactor mentioned in go/cast-details-view.

Bug: b/395947047
Flag: NONE refactor
Test: Cast to a device, click on the Cast tile, launched the dialog and
verified that the controller works.

Change-Id: Ic0b904908ff8c88e3557160d1970488daf5389d7

Former-commit-id: 69dff23684037d2ee106b2d0fa58e3ab6f76e0e9