[Flexiglass] Don't allow notification reordering while interacting with a HUN
When the user drags on a HUN to expand the shade, but within the same
gesture changes the swipe directon to snooze the HUN instead, we can get
into a temporarily state, where the HUN is already scheduled for removal
from the HeadsUpManager, but it still needs to be kept around.
The legacy stack solves this by NPVC reading HUM#isTrackingHeadsUp(),
and artifically keeping the shade expanded, until the user finishes the
swipe. This logic doesn't exist in Flexi, so this CL adds a new flag to
VisualStabilityCoordinator, which keeps isReorderingAllowed() false until
we have a tracked HUN.

Fixes: 391204020
Test: change swipe direction while snoozing a HUN
	- start swiping DOWN on a HUN (start opening the shade)
	- change the swipe direction to UP (decide to snooze instead)
        - Check: the HUN should be still displayed
Flag: com.android.systemui.scene_container
Change-Id: I5b138cd9c53ab86bc8a8f50d31c9b8331bd98546

Former-commit-id: 5042913aa9348ec7410fc921b356dd2ab79472af