WM callback to add/remove display in topology
This internal method allows Window Manager to inform Display Manager that a display should or should not be in the topology.

Remove the display group requirement in isDisplayAllowedInTopology() - in DisplayManagerServiceTest, external displays are initialized to not be in the default display group. Also, now we do want displays from multiple power groups to be in the topology.

Bug: 384013689
Flag: com.android.server.display.feature.flags.display_topology
Test: DisplayManagerServiceTest, DisplayTopologyCoordinatorTest
Change-Id: I4b1d63a4f15ed5eee48eec0889c8aaeb7621c989

Former-commit-id: 083631bfe0ca09573b74c07c7110dc741fb0296a