Don't show immersive confirmation for presentation
ag/31583844 made presentation windows immersive but as a side
effect, immersive confirmation started to show up.

This is undesirable for presentation windows because they are
designed so the user doesn't have to directly interract with the
window and focus should stay with their host activity.

This change adds window type argument to onImmersiveModeChanged()
so we can skip showing the dialog for presentations.

Flag: com.android.window.flags.enable_presentation_for_connected_displays
Bug: 397597400
Test: atest PresentationTest
Change-Id: I8e43711222584c2ec7efc31160d62ba03b936605

Former-commit-id: e2b9205b6447bc3132fbb8df6fe8efd29bba4b88