Prevent landscape bouncer from showing when opened from hub
As we enable rotation for hub, opening bouncer in landscape will show
it in a landscape layout before screen rotates to portrait.

This happens when KeyguardBouncerViewBinder#isShowing becomes true,
it starts the appear animation for bouncer elements right away. It
can happen earlier than hub -> bouncer transition start and screen
is still in landspe. The fix is to delay the animation and start it
after orientation changes to portrait. This can prevent a small
landscpe layout of bouncer from showing at the beginning.

Bug: b/392708288
Test: atest GlanceableHubToPrimaryBouncerTransitionViewModelTest
Test: atest KeyguardSecurityContainerControllerTest
Test: atest KeyguardSecurityContainerTest
Test: manually on foldable
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I7984dd312484e756abae3e34a96271b24c6ec0d5

Former-commit-id: 69f5eb75f2584c733ed08a62781e987abd3a968f