Move window-focus change subscription to WindowManagerInternal
WindowManagerService used to have a WindowChangeListener public
interface that allows clients to directly subscribe for focus and window
change events. In this change, we split the focus-change subscription
into a separate interface, and make it available via
WindowManagerInternal. This allows us to provide a better interface
between the service and other system-server components.

Bug: 397832722
Test: n/a mostly a refactor
Flag: EXEMPT trivial refactor
Change-Id: I1a91e409a8d2aa0541c91d6f5a5341f37e35390d

Former-commit-id: 63444192a19052997066b2071c340194b7ff32df