Fall back to checking window bounds height during drag-to-desktop
During the (spring) drag-to-desktop animation we fetch the current
animation fraction by checking the animated width of the window against
its start- and end bounds. This logic breaks if the start- and
end-widths are the same. In this CL we check the start- and end heights
if the widths are the same.

Bug: 397196997
Flag: com.android.window.flags.enable_desktop_windowing_enter_transition_bugfix
Test: manual
Change-Id: I2144503e3d56e44a55840bc750bc03478b58fe48

Former-commit-id: 65048ceb2f8d863f66853f9cb116190f6ee2f8d4